pull __version__ to top level
I pulled __version__ to the top level and made it a string literal.
Not sure if version_tuple is still needed, but I kept it.

setup-tools needs to get at __version__, and if it's not defined
to a literal, then setup-tools has to import the tinytuya module.

If it imports the module, then all the modules dependencies are needed
as build depdendencies. The motivation for this change is to avoid
that situation.

I pulled __author__ up to top level too just because it was there
next to __version__.