More linting, move __version__ back to core
I noticed I wasn't getting undefined variables, so I switched to
`extend-select` instead of just `select` to pick up the default
Ruff rules.

I then disabled some of the defaults to avoid making too many code
changes. I made a couple of minor changes to comply with the lints.

I realized core.py was actually using __version__ so I moved it
back to that place and restored the imports, and modified the
setuptools config to look for it there.